requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
requests==2.31.0
python-dotenv==1.0.0
pymongo==4.6.0
bcrypt==4.1.1
Werkzeug==3.1.4
